Update deps and adjust to PIA change

+++ Nodejs/Product/Nodejs/SharedProject/ProjectNode.cs
@@ -1266,7 +1266,7 @@ namespace Microsoft.VisualStudioTools.Project
             {
                 Array contextParamsAsArray = contextParams;
 
-                var result = ivsExtensibility.RunWizardFile(wizardToRun, (int)dlgOwner, ref contextParamsAsArray, out wizResultAsInt);
+                var result = ivsExtensibility.RunWizardFile(wizardToRun, dlgOwner, ref contextParamsAsArray, out wizResultAsInt);
 
                 if (!ErrorHandler.Succeeded(result) && result != VSConstants.OLE_E_PROMPTSAVECANCELLED)
                 {
